requirements.txt
tensorflow>=2.10.0
opencv-python>=4.6.0
geopandas>=0.12.0
matplotlib>=3.5.0
folium>=0.14.0
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0
rasterio>=1.3.0
shapely>=1.8.0
jupyter>=1.0.0
pillow>=9.0.0
tqdm>=4.64.0
seaborn>=0.11.0
plotly>=5.10.0 